requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
iniconfig==2.0.0
jax>=0.4.30
jaxlib>=0.4.30
kiwisolver==1.4.7
matplotlib==3.9.2
ml-dtypes==0.4.0
numpy==2.1.0
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
pytest==8.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.14.1
six==1.16.0
snakeviz==2.2.0
tornado==6.4.1
